parser/tests/tree_construction/basic.dat

#data
<!DOCTYPE html><math><mi>
#errors
(1,25) expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
